requirements.txt
Flask==2.2.5
Flask_Mailing==0.2.2
mysql_connector_repackaged==0.3.1
nltk==3.8.1
numpy==1.21.6
pandas==2.0.2
scikit_learn==1.2.2
tensorflow==2.11.0
tensorflow_intel==2.11.0
torch~=1.13.1

mysql-connector-python~=8.0.26
sklearn~=0.0.post1
scikit-learn~=1.2.2
Werkzeug~=2.2.2